Feat[#250] : CompareIndex
